refactor(codegen): extract states into plugin folder
