+++ scripts/codegen/src/schema.ts
@@ -16,6 +16,7 @@ import { exampleEntry } from "./plugins/examples/schema.ts";
 import { stateEntry } from "./plugins/latch/schema.ts";
 import { motionFragment } from "./plugins/motion/schema.ts";
 import { propEntry } from "./plugins/props/schema.ts";
+import { stateDef } from "./plugins/states/schema.ts";
 import { tokenEntry } from "./plugins/tokens/schema.ts";
 import {
   intentEntry,
@@ -51,23 +52,6 @@ const overlayBlock = z.strictObject({
   anchorVar: z.string().regex(/^--[A-Za-z0-9_-]+$/),
   mode: z.enum(["auto", "manual", "hint"]).default("manual"),
   modal: z.boolean().default(false),
-});
-
-// Shorthand `"open"` sugars to `{ to: "open" }`; long form unlocks `after:`,
-// `when:`, and `emits:`. Semantic checks reject shorthand when any of those
-// are needed.
-const transitionTarget = z.union([
-  z.string().min(1),
-  z.strictObject({
-    to: z.string().min(1),
-    after: z.string().min(1).optional(),
-    when: z.string().min(1).optional(),
-    emits: z.record(z.string(), z.record(z.string(), z.unknown())).optional(),
-  }),
-]);
-
-const stateDef = z.strictObject({
-  on: z.record(z.string(), transitionTarget).default({}),
 });
 
 // `fromChildren: true` makes the generator wrap the consumer's children in
